admin_service/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
